% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceLines.R
\name{sourceLines}
\alias{sourceLines}
\title{Run specific lines of a source file}
\usage{
sourceLines(sourcefile, sline = 1, fline = 1)
}
\arguments{
\item{sourcefile}{The path and file name of the source file.}

\item{sline}{The line to start reading from the source file. Default is 1.}

\item{fline}{The final line to be read in from the source file. Default is 1.}
}
\value{
None
}
\description{
This is a variation of the solution found on \href{https://stackoverflow.com/questions/26245554/execute-a-set-of-lines-from-another-r-file}{Stack Overflow}.
The function will take a range of lines within a source file and execute only those lines, via the \code{source()} function.
}
